mcp-server-gitparser/pyproject.toml
[project]
name = "mcp-server-gitparser"
version = "0.1.0"
description = "MCP Server for parsing GitBook sites and GitHub repositories into Markdown"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.8.0,<2.0.0",
    "fastmcp>=2.3.4",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.5.0",
    "gitingest>=0.3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server_gitparser"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.2",
    "ruff>=0.1.6",
    "black>=23.11.0",
    "isort>=5.12.0",
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "W",   # pycodestyle warnings
    "C90", # mccabe complexity
    "N",   # pep8-naming
    "D",   # pydocstyle
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "S",   # flake8-bandit
]
ignore = [
    "D100", # Missing docstring in public module
    "D104", # Missing docstring in public package
    "D107", # Missing docstring in __init__
    "D203", # 1 blank line required before class docstring
    "D212", # Multi-line docstring summary should start at the first line
    "S101", # Use of assert detected (common in tests)
]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "build",
    "dist",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--disable-warnings",
    "-v",
]
